+++ core/docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+
+# 1.15.5
 ### Features
 * Added a new method `WaitForEndpointAvailability` to the API container which can be used to wait for a service's HTTP endpoint to come up
 * Added a new method `GetServiceIP` in service network which returns an IP Address by Service ID.